menambahkan arsitektur multi talent

+++ app/Http/Controllers/Admin/EventController.php
@@ -10,11 +10,25 @@ use Illuminate\Support\Facades\Storage;
 
 class EventController extends Controller
 {
-    public function index()
-    {
-        $events = Event::with('category')->latest()->paginate(10);
-        return view('admin.events.index', compact('events'));
+  public function index()
+{
+    if (auth()->user()->role == 'superadmin') {
+
+        $events = Event::with(['category', 'organization'])
+            ->latest()
+            ->paginate(10);
+
+    } else {
+
+        $events = Event::with(['category', 'organization'])
+            ->where('organization_id', auth()->user()->organization_id)
+            ->latest()
+            ->paginate(10);
+
     }
+
+    return view('admin.events.index', compact('events'));
+}
 
     public function create()
     {
@@ -40,6 +54,7 @@ class EventController extends Controller
                 ->store('posters', 'public');
         }
 
+        $data['organization_id'] = auth()->user()->organization_id;
         Event::create($data);
 
         return redirect()
@@ -53,14 +68,30 @@ class EventController extends Controller
     }
 
     public function edit(Event $event)
-    {
-        $categories = Category::all();
-
-        return view('admin.events.edits', compact('event', 'categories'));
+{
+    if (
+        auth()->user()->role != 'superadmin'
+        &&
+        $event->organization_id != auth()->user()->organization_id
+    ) {
+        abort(403);
     }
+
+    $categories = Category::all();
+
+    return view('admin.events.edits', compact('event', 'categories'));
+}
 
     public function update(Request $request, Event $event)
     {
+         // Cek apakah organizer berhak mengedit event ini
+    if (
+        auth()->user()->role != 'superadmin' &&
+        $event->organization_id != auth()->user()->organization_id
+    ) {
+        abort(403);
+    }
+    
         $data = $request->validate([
             'category_id' => 'required|exists:categories,id',
             'title' => 'required|string|max:255',
@@ -90,15 +121,23 @@ class EventController extends Controller
     }
 
     public function destroy(Event $event)
-    {
-        if ($event->poster_path) {
-            Storage::disk('public')->delete($event->poster_path);
-        }
-
-        $event->delete();
-
-        return redirect()
-            ->route('admin.events.index')
-            ->with('success', 'Data event berhasil dihapus secara permanen.');
+{
+    if (
+        auth()->user()->role != 'superadmin'
+        &&
+        $event->organization_id != auth()->user()->organization_id
+    ) {
+        abort(403);
     }
+
+    if ($event->poster_path) {
+        Storage::disk('public')->delete($event->poster_path);
+    }
+
+    $event->delete();
+
+    return redirect()
+        ->route('admin.events.index')
+        ->with('success', 'Data event berhasil dihapus secara permanen.');
+}
 }
